ai-service/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
pillow           # Для работы с изображениями
numpy